Fixed serialization and updating added dynamic updated questions, added setters and getters so that spring finds them and use them correctly when parsing questions json. Given that long ids needs to be generated by server-side, client side will send questions without ID, those will be then generated server-side sequentially for optimization and future fetching purposes and returned to the client in the response body for future api calls

+++ server/src/main/java/nl/tudelft/oopp/livechat/controllers/QuestionController.java
@@ -25,7 +25,7 @@ public class QuestionController {
     * This endpoint will be used to ask a question
     */
     @PostMapping("/ask")
-    public int askQuestion(@RequestBody QuestionEntity question) {
+    public long askQuestion(@RequestBody QuestionEntity question) {
         return questionService.newQuestionEntity(question);
     }
